feat(145409): Refatora modal do históricos de habilitados para tela
